Fixed a few tests and the way inheriting increment strategy works
